--- src/main/java/or/sopt/houme/domain/taste/dto/response/MoodBoardListResponse.java
@@ -1,11 +1,0 @@
-package or.sopt.houme.domain.taste.dto.response;
-
-import java.util.List;
-
-public record MoodBoardListResponse(
-        List<MoodBoardResponse> moodBoardResponseList
-) {
-    public static MoodBoardListResponse of(List<MoodBoardResponse> moodBoardResponseList) {
-        return new MoodBoardListResponse(moodBoardResponseList);
-    }
-}
