--- src/main/java/or/sopt/houme/domain/taste/entity/TasteTag.java
@@ -1,32 +1,0 @@
-package or.sopt.houme.domain.taste.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@AllArgsConstructor
-@Builder
-@Table(name = "taste_tags")
-public class TasteTag {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "taste_id")
-    private Taste taste;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "tag_id")
-    private Tag tag;
-
-
-    public static TasteTag of(Taste taste, Tag tag) {
-        return TasteTag.builder()
-                .taste(taste)
-                .tag(tag)
-                .build();
-    }
-}
