--- src/main/java/or/sopt/houme/domain/taste/service/TagService.java
@@ -1,12 +1,0 @@
-package or.sopt.houme.domain.taste.service;
-
-import or.sopt.houme.domain.taste.entity.Tag;
-
-public interface TagService {
-
-    // userId와 imageId로 1순위 태그 찾기
-    Tag findTagByUserIdAndImageId(Long userId, Long imageId);
-
-    // 무드보드 Id (tasteId)로 Tag 찾기
-    Tag findTagByTasteId(Long tasteId);
-}
