--- src/main/java/or/sopt/houme/domain/taste/service/TasteService.java
@@ -1,15 +1,0 @@
-package or.sopt.houme.domain.taste.service;
-
-import or.sopt.houme.domain.taste.dto.response.MoodBoardListResponse;
-import or.sopt.houme.domain.taste.entity.Taste;
-
-import java.util.List;
-
-public interface TasteService {
-
-    // 무드 보드 이미지 제공 API
-    MoodBoardListResponse getMoodboard(Long cursorId, int size);
-
-    // 무드 보드 객체 제공
-    List<Taste> getTasteList(List<Long> tasteList);
-}
